Update `rz_phase_gradient` docs (#8536)
**Context:**
Phase gradients come with some conventions about the sign of the phase
and the wire ordering

**Description of the Change:**
Update the docs to have the "generic" sign from the literature (afaik)
and the PL wire ordering.
Also simplify the state prep of phase gradient states in tests and docs.

**Benefits:**
Updated docs, unification with soon-to-exist compilation hub entry.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-99668]