[Capture] Prepare grad prim for catalyst translation (#8357)
**Context:**

When lowering to mlir, catalyst needs to know name and hash of the
original function to properly name the `FuncOp` and to cache the created
`FuncOp`, respectively.

See https://github.com/PennyLaneAI/catalyst/pull/2078 for the
corresponding catalyst PR.

So we need to pass on the `fn` to the lowering.

**Description of the Change:**

* Stores `fn` as a piece of metadata in `grad_p`.

* Removed `jacobian_p` in favor of a single `grad_p` with a `scalar_out`
kwarg. Less code to maintain.

**Benefits:**

We can define lowerings for the primitives in catalyst.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-100398]

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>
Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>