Improve gate_set and stopping_condition resolution in decompose (#8532)
**Context:**

**Description of the Change:**

- `gate_set` is an iterable of operator types and/or operator names, or
a dictionary mapping operator types to their respective costs.
- The `stopping_condition` is a callable that returns `True` iff an
operator instance does not need to be further decomposed.
- The final stopping condition is a combination of `gate_set` and
`stopping_condition` (we stop decomposing if the gate meets the stopping
condition, or if the gate is part of the gate set)

What if the `gate_set` isn't provided?
- If both `gate_set` and `stopping_condition` are not provided, we treat
`qml.ops.__all__` as the gate set. Essentially, we would only decompose
templates and user-defined custom operators.
- If `stopping_condition` is provided but the `gate_set` isn't provided,
an error is raised if graph is enabled, because if a stopping condition
is provided, we assume that the user does actually want to decompose
beyond templates, and in this case, a `gate_set` must be explicitly
provided to the graph.

Graph Enabled:
- `gate_set` is required, `stopping_condition` is optional

Graph Disabled:
- Either specify `gate_set`, or `stopping_condition`, or both (used to
raise an error when both are specified).
- If `gate_set` is a dictionary mapping gate types to weights, ignore
the weights, and treat the keys as the gate set (used to raise an error)

**Benefits:**

More coherent approach to resolving the `gate_set` and
`stopping_condition`.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-102068]

---------

Co-authored-by: Marcus Edwards <marcusstephenedwards@gmail.com>
Co-authored-by: Marcus Edwards <94946848+comp-phys-marc@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>