Fix QASM output for circuits with mid-circuit measurements (#8556)
**Context:** In trying to upgrade mitiq to the latest version of
pennylane, we discovered that QASM output from pennylane contains code
that doesn't get properly parsed by cirq's parser. It's possible this is
a bug in cirq, but I'm 90% sure the missing semicolon is unintended.

This bug is so famous, it's almost hard to believe I came across one
IRL.

**Description of the Change:** Ensure `creg mcms[...]` is terminated
with a semicolon in QASM output.

**Benefits:** QASM output will be parsed by cirq qasm parser (our use
case).

**Possible Drawbacks:** low risk: downstream code comparing exact qasm
strings that relied on the missing semicolon may need updates.

**Related GitHub Issues:** No open tickets, but building on work started
in https://github.com/PennyLaneAI/pennylane/pull/8210


**Question:** The pennylane circuit I started with to generate the qasm
doesn't have mid-circuit measurements, but only one terminal
measurement. Is it surprising that a mid-circuit measurement was created
from it?

---------

Co-authored-by: Marcus Edwards <94946848+comp-phys-marc@users.noreply.github.com>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>