Remove capture cache on qnode (#8629)
**Context:**

The capture cache on the qnode is causing maintainence issues, and is no
longer needed as capture is just a frontend for catalyst.

**Description of the Change:**

Remove everything to do with caching jaxpr at the qnode level.

**Benefits:**

Unblocks bumping jax version and proper handling of keyword arguments.

**Possible Drawbacks:**

If we ever decide that we actually want this again, we have to add it
back in.

**Related GitHub Issues:**

[sc-103737]

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>