Control flow fallback (#8615)
**Context:**

One of the challenges with turning autograph on by default is that it
converts things it shouldn't convert into a structured for loop. When
that happens, the whole program just fails and falls over.

To make things more user friendly, we instead just fall back to a
standard python for loop. Less performant, but at least it works.

**Description of the Change:**

Fall back to standard python for loop if the capture of the jaxpr fails.

**Benefits:**

Things continue to run.

This fallback can be turned off by turning the
`qml.exceptions.CaptureWarning` into an error.

**Possible Drawbacks:**

Things are unrolled, even if unrolled with a warning.

**Related GitHub Issues:**

[sc-103710]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>