Deprecate callable gate set (#8533)
**Context:**
The `decompose` transform accepts both a `gate_set` and a `stopping
condition`. The `gate_set` is supposed to be a static iterable of
operator types and/or operator names, and the `stopping_condition` is
supposed to be a function that returns `True` or `False` given an
operator instance. For backwards compatibility, we still accepts a
function as the `gate_set`, but we internally treat it as the
`stopping_condition`. This is confusing, and we should be more strict
about what exactly is expected for each argument.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-102183]
Blocked By: https://github.com/PennyLaneAI/pennylane-lightning/pull/1290

---------

Co-authored-by: Marcus Edwards <marcusstephenedwards@gmail.com>