[Capture] switch to a single transform primitive (#8576)
**Context:**

As a step 1 for making it so that we have only one type of transform
system between pennylane and catalyst and as part of generally making
capture a bit more friendly, we need to have one single transform
primitive.

This makes it so we don't have to keep registering every single
transform is a whole bunch of places. We only have to pay attention to
the single transform primitive.

**Description of the Change:**

Switches to having a single transform primitive that takes a `transform`
kwarg.

**Benefits:**

Easier to deal with transforms.

**Possible Drawbacks:**

Breaking change, but breaking change for internal detail.

**Related GitHub Issues:**

[sc-103382]


Note that this change breaks catalyst, so we will have to override merge
this PR in, then merge https://github.com/PennyLaneAI/catalyst/pull/2177
immediately after.

---------

Co-authored-by: Mudit Pandey <18223836+mudit2812@users.noreply.github.com>