[FTQC][Unified Compiler] Outline MBQC transform (#8619)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

[sc-103563]

The MBQC transform has been outlined in the subroutines in the
`convert_to_mbqc_formalism` pass. This would reduce the number of lines
in the IR if the circuit is unrolled.

A few other improvements/considerations:
- Reduced the number of arith.ConstantOp used in defining the
measurements.
- The `visibility` of subroutines are set as `private` to work better
with the `symbol-dce` pass defined in the catalyst pipeline. Note that
current pass inserts subroutines for all gates in the MBQC gate set.
This could be redundant if not all gates are applied in a given circuit.
However, there is no need to worry about that since the `symbol-dce`
pass would eliminate the unreferenced subroutine given that its
`visibility` is private.
- Note that those subroutines are attached with an attribute
`mbqc_transform`. This attribute is used to skip the
`convert_to_mbqc_formalism` transform for the subroutine (There are
"Hadamard" gates in the subroutine, which we don't want to replace it
with a subroutine call).
- Note that `IfOp` in the subroutines could be hitted by the
`convert_to_mbqc_formalism` transform. No need to worry for now as there
are no gates in the MBQC gate set there.
- Note that the order of parameters for the subroutine is aligned with
the corresponding qml.operation.

**Description of the Change:**

**Benefits:**

Take the `ncas = 10` (or `21` qubits) XAS circuit ( `stateprep`
ignored), the number of lines of `circuit.mlir` is `26158`, compared to
`92185` for the `master` branch. The number of lines of `_llvm.ir.ll` is
`213173`, compared to `420280` for the `master` branch.

**Possible Drawbacks:**

**Related GitHub Issues:**