Fixing tag refs for Release Automation (#8678)
**Context:**
The last step release automation has a bug that [prevents it from
uploading the tag for the rc
branch](https://github.com/PennyLaneAI/pennylane/actions/runs/19480175172/job/55750110025#step:3:12)
because the tag is the same name as the branch.

**Description of the Change:**
Specifying `tag` when pushing the branch tag.

**Benefits:**
Branch tag gets pushed.

**Possible Drawbacks:**
None.

**Related GitHub Issues:**
https://github.com/PennyLaneAI/pennylane/pull/8352