Add skip_decomp_matrix_check to assert_valid (#8687)
**Context:**

Some decomposition rules produce circuits that we cannot obtain the
matrix of (e.g., decomposition rules that contain mid-circuit
measurements). When we use `assert_valid` on these decomposition rules,
we cannot check that the matrix is correct, but we still want to verify
other things (such as that the decomposition produced gates that match
the resource function)

**Description of the Change:**

Added an `skip_decomp_matrix_check` argument to `assert_valid` that
allows the test to skip the matrix check part of testing a decomposition
rule but still verify that the resource function is correct.

**Benefits:**

More fine-grained control over decomposition rule verification.

**Possible Drawbacks:**

**Related GitHub Issues:**
Unblocks work for https://github.com/PennyLaneAI/pennylane/pull/8633