Fallback to graph decomp if compute_decomposition not defined (#8686)
**Context:**

When reviewing the introduction of a new template, I realized it's
really quite annoying that we are needing to add decompositions in two
different ways for new operators. We should be able to just define the
decomposition once and have things work throughout or ecosystem.

This will make it much easier to shift back and forth between the two
systems with a minimum of pain.

**Description of the Change:**

Has `Operator.decomposition` fallback to the first registered graph
decomp.

**Benefits:**

Easier to add new operators and have them compatible with everything.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-104505