Quantum automatic differentiation (#8640)
**Context:** Implementing the quantum automatic differentiation
algorithm detailed in [this paper ](https://arxiv.org/pdf/2408.05406)
from the [Quantum Systems Lab](https://csl.yale.edu/) at Yale
University.

**Description of the Change:**

1) We check if we have a work wire -> direct or standard
differentiation.

2) check if we are doing forward or reversed (switch the generator with
the measured op) based how many combinations there are. Depends on the
combinations of expectations and terms in the Hamiltonian, for example.
i.e. if the terms in the Hamiltonian are on different wires we can
measure them at the same time, but still need different circuits for the
controls, can refer to the hamiltonian’s "grouping indices" list: its
length gives the number of shots need for the expectations. See Table
III in https://arxiv.org/pdf/2408.05406 for exact formulas.

**Benefits:** Allows the optimal differentation method to be selected
automatially based on the number of shots required and the availability
of auxilliary wires.

**Possible Drawbacks:** Still a Work In Progress (WIP) needing feeback!

**Related ShortCut Stories:** [sc-103833]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>