Compact Hamiltonian classes are modified to accept one-norm (#8697)
**Context:**
Compact Hamiltonian classes need 1-norm to get full resource estimates
for some algorithms. It is now added to be an optional argument.

**Description of the Change:**
Compact Hamiltonian classes now accept 1-norm as an optional argument. 
Typechecks are added to all the classes to avoid passing of incorrect
data types.
Unit tests are added for the compact Hamiltonian classes.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-103102]

---------

Co-authored-by: Jay Soni <jbsoni@uwaterloo.ca>
Co-authored-by: Austin Huang <austin.huang@xanadu.ai>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>