chore: add drop of numpy 1.X support messaging (#8578)
We are dropping 1.X maintenance support and so we want to let the user
know this.

```
❯ uv pip install numpy~=1.26
Resolved 1 package in 58ms
Uninstalled 1 package in 191ms
Installed 1 package in 31ms
 - numpy==2.3.4
 + numpy==1.26.4
❯ python
Python 3.12.8 (main, Dec 19 2024, 14:22:58) [Clang 18.1.8 ] on darwin
Type "help", "copyright", "credits" or "license" for more information.
>>> import pennylane as qml
../pennylane/pennylane/__init__.py:220: PennyLaneDeprecationWarning: PennyLane v0.44 has dropped maintainence support for NumPy < 2.0.0. You have version 1.26.4 installed. Future versions of PennyLane are not guaranteed to work with NumPy<2.0. Please consider upgrading NumPy using `python -m pip install numpy --upgrade`.
  warnings.warn(

```

Eventually down the road we will need to require numpy>=2.0.0.

[sc-101547]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Diego <67476785+DSGuala@users.noreply.github.com>