BugFix: ChangeOpBasis._flatten operator order (#8722)
**Context:**

`ChangeOpBasis` internally reverses the order of operations passed in,
so when `_flatten` is called, it needs to return the leaves in the
reverse order so that when `_unflatten` is called, the same operator is
reconstructed.

**Description of the Change:**

Additionally, the program structure of the decomposition of
`TemporaryAND` is simplified.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
Fixes https://github.com/PennyLaneAI/pennylane/issues/8721
[sc-105069]