BugFix: DecompositionGraph incorrectly exclude rules (#8725)
**Context:**

The decomposition graph reuses operator nodes that are not work-wire
dependent, even under a different work wire budget. For example, if
`CustomOp` dynamically requests a work wire to decompose to `SimpleOp`,
then in graph construction, the node for `SimpleOp` will be labeled with
a work wire budget of `total number of work wires - 1`. However, if the
decomposition of `SimpleOp` does not use work wires, it should not be
treated as "infeasible" even if the total number of work wires is 0,
making the work wire budget for `SimpleOp` technically negative.

**Description of the Change:**

Adds a check to `is_feasible` and returns `True` if the decomposition
does not depend on work wires.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
Fixes https://github.com/PennyLaneAI/pennylane/issues/8720
[sc-105051]