Fix Doctest Ci after jax bump (#8724)
-------------------------------------------------------------------------------------------

**Context:**
After bump jax to 0.7.0 and 0.7.1, some doctests are affected. We fix
them here

**Description of the Change:**
1. `transform.py` the expected output tracing info need to be adjusted
as the hashable version
2. `condition.py` contains some dynamic shape functionalities, which
should be skipped for now, until the [patch
PR](https://github.com/PennyLaneAI/pennylane/pull/8525/) is merged.

**Benefits:**
Documentation Tests should be back to normal and almost always green now
(except for the mysterious `qmc.py` which we have no clue yet)

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
[sc-105091]