Transform program dispatch all (#8731)
**Context:**

Add ability to dispatch `TransformProgram` to anything individual
transforms can dispatch onto.

Right now a TransformProgram can be applied to only batches of tapes and
ClosedJaxpr.

For any other object, we should just chain apply the contents of the
TransformProgram to the input.

So after this, TransformProgram will be able to be applied to devices
and qnodes.


**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-103781]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>