Upgrade resources of `PauliRot` to match Pennylane (#8562)
**Context:**
While developing the demo `How to: Resource estimation in PennyLane`. We
discovered some mismatch between the resources used in pennylane
decompositions compared to the resource estimates from the estimator
module. In this PR we introduce two main changes to the estimator module
to provide more accurate (and efficient) estimates for the
`TrotterProduct` and `PauliRot` templates.

**Description of the Change:**
- Improved the mapping functionality between `Operator` and
`ResourceOperator` to allow for mapping instances of the
`TrotterProduct` class where the underlying hamiltonian had some
commuting group structure. In practice, this means users can now map a
hamiltonian that is encoded as a nested `Sum` of `Sum` operators and
obtain accurate resource estimates.
- Added two special cases to the decomposition of `PauliRot` when the
input Pauli string is `"XX"` or `"YY"` to use the more optimal IsingXX,
IsingYY decompositions.
 
**Benefits:**
- Resource estimates are reduced and more accurate 
- Resource estimates match results from PennyLane 

**Possible Drawbacks:**
None

**Related GitHub Issues:**
None

---------

Co-authored-by: Austin Huang <austin.huang@xanadu.ai>
Co-authored-by: ANT0N <39093564+AntonNI8@users.noreply.github.com>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Soran Jahangiri <40344468+soranjh@users.noreply.github.com>
Co-authored-by: Diego <67476785+DSGuala@users.noreply.github.com>