Rename TransformContainer to BoundTransform (#8753)
**Context:**

We are making the transforms infrastructure more user friendly and user
facing, so we have a clear, intuitive way to construct compile pipelines
for reuse.

One of the current components for our transforms infrastructure is
`TransformContainer`. But it has a rather unfortunate, unintuitive name.
To make it the name more obvious, I'm proposing we rename it
`BoundTransform` *before* we start exposing it to more people

**Description of the Change:**

Find and replace `TransformContainer` -> `BoundTransform`.  

Add docs to `BoundTransform`.

Leave `TransformContainer` as an alias.

**Benefits:**

Everything flows so much more naturally!

**Possible Drawbacks:**

We are renaming things. And we are just leaving the historical name
around as a remnant.

**Related GitHub Issues:**

[sc-105453]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>