Some quality of life improvements for `estimator` module.  (#8761)
**Context:**
While writing demos, we came across of a few small changes that help
improve the user experience when working with the `estimator` module

**Description of the Change:**
- The default and controlled resources of `QROM` were improved in the
edge case where `restored = True` and `sel_swap_depth = 1`.
Specifically, in this edge case we don't need to the expensive
uncomputations that we would otherwise do when `restored = True` because
we don't actually have any additional swap registers to restore.
- `res.total_gates` and `res.total_wires` properties were added to the
`Resources` class to make it easier for users to access those quantities
directly.
- `pauli_ham.num_terms` property was added to make it easier for users
to access the total number of pauli words of their hamiltonian.

**Benefits:**
Resources for a common subroutine are reduced and certain important
properties of the Resources object or the PauliHamiltonian object are
easier to access.

**Possible Drawbacks:**
None

**Related GitHub Issues:**
[[sc-105806](https://app.shortcut.com/xanaduai/story/105806)]

---------

Co-authored-by: Diksha Dhawan <40900030+ddhawan11@users.noreply.github.com>
Co-authored-by: Soran Jahangiri <40344468+soranjh@users.noreply.github.com>
Co-authored-by: Diego <67476785+DSGuala@users.noreply.github.com>