Fix the bug in QubitizeTHC where the auxiliary wires are not being accounted for accurately. (#8719)
**Context:**
There was a bug in SelectTHC, where the auxiliary wires for QROMs in
SelectTHC template were not being accounted for accurately when
Select-SWAP variant of QROM was being used. This PR fixes that bug.
There is also some refactoring done to make sure algorithmic wires for
PrepTHC are being counted accurately when it is used inside the
QubitizeTHC template.

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Austin Huang <austin.huang@xanadu.ai>
Co-authored-by: Jay Soni <jbsoni@uwaterloo.ca>
Co-authored-by: Soran Jahangiri <40344468+soranjh@users.noreply.github.com>