soften import of networkx (#8737)
**Context:**

Python 3.14.1 has an issue that broke the import of `networkx`. To make
it so we can at least import pennylane, we are softening the import of
networkx to only occur locally when needed, instead of globally. While
the relevant features like qcut will not be a available on 3.14.1,
everything else should be perfectly fine now.

**Description of the Change:**

Networkx only imported on type checking and when accessing relevant
functionality.


**Benefits:**

Almost all functionality exists with Python 3.14.1.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-105446]