[Capture] `qml.cond` adds an empty return if function partial of operator (#8776)
**Context:**

Now that `qml.cond` treats all inputs as potentially trainable, we need
to use partials to handle static metadata.

But when using a partial, we skip over the special logic for a true
function that is an operator type. The true fn itself isn't an operator
type, just the partial of an operator type.

**Description of the Change:**

Add in the special return handling for a true fn that is either an
operator type, or a **partial** of an operator type.

**Benefits:**

We can do `qml.cond(m, partial(qml.PauliRot,
pauli_word=pauli_word))(theta, wires=wires)` without error.

**Possible Drawbacks:**

**Related GitHub Issues:**