Integrate qml.measure with catalyst.measure (#8782)
**Context:**

It's been taking way too long to finally make `qml.measure` dispatch to
`catalyst.measure` when needed, and causing me a lot of headaches for
integrating everything with program capture.

So I decided to stop waiting and just make it happen.

This will make it *substantially* easier to port everything to using
program capture as the frontend, because we will be able to use the same
`measure` function in both cases.

**Description of the Change:**

`qml.measure` dispatches to `catalyst.measure` when necessary.

**Benefits:**

My level of frustration will drop substantially.

We can use the same workflows for both catalyst frontends.

**Possible Drawbacks:**

Something something qjit and tape-pennylane support different
`mcm_methods`?

**Related GitHub Issues:**

[sc-106599]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>