Remove ResourcesUndefinedError in symbolic ops (duplicate) (#8811)
**THIS IS A DUPLICATE OF #8598 DUE TO A BUG IN GITHUB WRT UNRESOLVED
CONVERSATIONS**

**Context:**
The `ResourcesUndefinedError` is used as control flow to handle cases
where a symbolic resource decomp method is not implemented on a resource
operator. It is unnecessary because we can simply use the inheritance
pattern to handle when a ResourceOperator does not implement a resource
decomp method.

**Description of the Change:**
We remove the `ResourcesUndefinedError` and the related try-except
blocks and move the functionality of the except-block into the
ResourceOperator class itself. This allows inheritance to take care of
the original functionality.

**Benefits:**
Cleans up the code, more extensible and maintainable and allows for
future improvements.

**Possible Drawbacks:**
The default symbolic resource decomp methods no longer live on the
Adjoint, Controlled, and Pow classes. This could be confusing.

**Related GitHub Issues:**
[sc-99796]

---------

Co-authored-by: Diksha Dhawan <40900030+ddhawan11@users.noreply.github.com>
Co-authored-by: Jay Soni <jbsoni@uwaterloo.ca>